global-streetscapes/requirements-cv-linux.txt
pip==24.0
setuptools==69.5.1

aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
cmake==3.26.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.10
GitPython==3.1.34
gpustat==1.1.1
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.12.2
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
lightning-utilities==0.9.0
lit==15.0.7
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.26.4
#nvidia-cublas-cu11==11.10.3.66
#nvidia-cuda-cupti-cu11==11.7.101
#nvidia-cuda-nvrtc-cu11==11.7.99
#nvidia-cuda-runtime-cu11==11.7.99
#nvidia-cudnn-cu11==8.5.0.96
#nvidia-cufft-cu11==10.9.0.58
#nvidia-curand-cu11==10.2.10.91
#nvidia-cusolver-cu11==11.4.0.1
#nvidia-cusparse-cu11==11.7.4.91
#nvidia-ml-py==12.535.133
#nvidia-nccl-cu11==2.19.3 #2.14.3
#nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
overrides==7.4.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.8
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
scikit-learn==1.3.0
scipy==1.12.0
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.1
torchaudio==2.2.1
torchmetrics==1.1.1
torchvision==0.17.1
tornado==6.3.3
tqdm==4.66.1
# traitlets==5.9.0
# triton==2.2.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wandb==0.15.9
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
yarl==1.9.2
zipp==3.17.0
Bottleneck==1.3.7
Brotli==1.0.9
gmpy2==2.1.2
huggingface-hub==0.20.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
numexpr==2.8.7
PySocks==1.7.1
regex==2023.10.3
safetensors==0.4.2
tokenizers==0.15.1
transformers==4.37.2
wheel==0.41.2